ch4/am: Only reference RMA target datatypes when necessary
In generic RMA, origin processes were adding/releasing references to
target datatypes unconditionally. Since the target datatype is
immediately flattened and sent to the target, the only instance where a
reference is needed is when the flattened type is large enough that we
do not copy it into the message header. In that case, the flattened
buffer lives in the MPIR_Datatype struct, so a reference must be added
to keep it from being freed if the user frees the datatype. The
reference can then be released when the flattened send is locally
complete.