test: add barrier in errors/pt2pt/truncmsg1.c
In this test the sender sends a huge message while the receiver posts a
tiny buffer. The receiver will quickly complete on receiving the first
packet, acknowledge and exit, while the send may still be sending the
remaining packets. The sender may fail due to disconnection. The
behavior is system dependent. On linux, the send is often able to
complete before detecting the disconnect. On FreeBSD, this test nearly
always fails with I/O error at the libfabric sockets layer.

Because the purpose of test is not about disconnection error, and the
disconnection error can be legitimate. Therefore, we are adding a barrier
to prevent the disconnect error. The test still tests the
implementations handling of truncation error.