configure: skip configure submodules if already built
When using embedded modules, check whether the corresponding convenience
library already exists and skip the configure if it does.

Note: the prebuilt module Makefiles contains absolute paths that may be
different from the current working dir. Thus we have to remove the
module subdir from the Makefile as well. This works as long as we do not
need to re-configure the pre-built submodules. If we ever need to
re-configure submodules, simply remove the la files, e.g.
  find modules '*.la' | xargs rm
and rerun configure.