ch4/ofi: protect mr_key_allocator in per-vci thread-cs
We need protect the global mr_key_allocator. This issue manifests in
when we --disable-ofi-domain, i.e. using SEP context for vci since
collided keys are being used in the single fi domain. However, even with
using fi domain for vci, we are still corrupting the global data, which
need to be fixed.

This is in the sending huge path when we using local provider key. The
performance impact of the lock should be small.