romio: use constant tag in collective write
The code was using `myrank + i + 100 * iter` for tag. With large data or
large number of processes, this tag may exceed the tag limit. We find
this complex tag is unnecessary. Because the message is send and
received using explicit source and destination, we don't need the rank
info in the tag. Because MPI message matching is guaranteed to be in
order, we don't need iteration number in the tag either. Thus, simply
using a constant tag is suffice for the correctness.