mpidu/genq: Add param to specify genq pool
Allow the caller of the function to get a cell out of a genq to specify
which process's pool to grab the cell from. Currently, the only user of
this function is the iqueue code so for now, it will keep the current
behavior of getting a cell from the sending process.