mpidu/genq: Use first-touch for memory binding
Force genq cells to be placed in the memory closest to the processes
to which they are assigned by memsetting them to 0. This will ensure
that when copying data into them, they won't be migrated to the process
that is writing the data and will instead stay with the process that
will read the memory.