github: Add CODEOWNERS file
Add initial CODEOWNERS file that sets @pmodels/mpich-intel as owner for
collective frameworks initially contributed by the Intel team. Pull
requests to this code, or any code with an owner specified, will
automatically request a review from said owner(s).

The overall intention is for code owners to have the authority to accept
or reject code changes based on whatever criteria they prefer, with some
exceptions. For example, security or licensing issues may result in
changes to the code by the core MPICH team without involvement of the
code owner.