autogen: let -do=xxx execute a single step function
Mixing the --with-xxx and -do=xxx options makes the script logic very
complex since there are extensible steps and many implicit dependency
conditions. Instead, let -do=xxx just execute a single step, which I
believe it is the original intention of this option. We assume that when
this option is used, the developer knows the dependency are already done
or checked.