configure: allow work-around when Python 3 is unavailable
We believe it is uncommon to have a system without Python 3. But in case
it is unavailable, one still can manually run e.g. `python3
maint/gen_binding_f08.py` separately to proceed, with the caveat that
they need know the options if the default doesn't work.

Also rename the variable "cmd" into "cmd_f08" to prevent collision with
other parts of  configure.