test: fix fortran tests with MPI_IN_PLACE
Even though we ignores send counts and displacements when MPI_IN_PLACE
is used, fortran still checks the argument types. While we could add
directives to disable the check for some compilers, it is not too much
trouble for users to use dummy arrays, and I think it is useful not to
disable this compiler checks.