mpl: remove AI_V4MAPPED from sockaddr
The purpose of AI_V4MAPPED flag is to be able to use IPv6 when hosts are
only configured with IPv4 addresses. Unfortunately, AI_V4MAPPED is not
widely supported, e.g. on OpenBSD. And since the use scenario is not
common -- currently we default to IPv4 -- it is safe to remove this
hint.