ch4/posix: force a 16-byte alignment
On 32-bit systems the `MAX_ALIGNMENT` is 4. Breaking the segments at
boundary of 4 may endup chopping a `double` in half. Currently, yaksa
cannot handle packing or unpacking fragments of basic types.

This patch forces a 16-byte minimum payload alignment. It will prevent
most fail cases but it is not fail proof. For example, if we are packing
`int+double`, neither alignment of `4` or `16` can prevent splitting a
`double`. Yaksa need handle the fragments for a complete fix.