ch3/win: Stash and ignore optimized MR hint
A new info hint was added in 3cbc32e9ba285c29d2063ff8b65778597d14b1ee to
support optimized memory regions in CH4, but not putting it in CH3
caused problems with tests that look for it.

This patch adds that hint to CH3 (even though it is never used) to allow
the test to pass.