runtests: run multiple tests using run_mpitests
Filter the test list and find all tests that can be run using
run_mpitests and run them first. Tests are grouped by testlist and
number of processes.

The running of tests are controlled by runtests using the input/output
pipes, thus we still have the granular control of individual tests.

When run_mpitests abort due to e.g. segfault, restart it with next test
in the testlist. Track the number of such restart and abort in case
something systematic causing it to fail repeatedly, for example, error
in run_mpitests itself.