ch4: remove incorrect resetting of VCI hints
VCI hints were being initialized to default values at an incorrect
and unnecessary location. This was leading to wiping out the earlier
set VCI hints. As a result, info hints were not applied when passed
to the comm creation routines such as MPI_Comm_dup_with_info.

This patch fixes this problem.