debugger: Rename debug queue entry pointer
Send requests contain a pointer to a parallel entry in the send queue
debugging list. It is misleading to call this entry "dbg_next" since it
refers to the request itself. Note that if the send queue debugging code
were to search for the request in the debug list, this field would not
be necessary to have in the MPIR_Request.