mpidu/genq: Use the new MPMC queue for free queues
This is necessary in order to allow the free queues to be used for
either the sender or the receiver.  For receiver-side free queues, this
could have been accomplished with an SPMC queue, but implementing that
has a very similar cost as an MPMC queue, so we'll use that instead.
Having the MPMC queue also gives us the most flexibility.