op: replace predefined datatype on input
We are supposed to support operations on some predefined datatype, such
as those from MPI_Type_create_f90_xxx. However, those types are in fact
derived types wraps around a basic type. Check and replace the input
datatype with the basic type if that is the case so the reduction can
work. Since the datatype in these case are all input only parameters,
there shouldn't be any side effects replacing them.