coll: fix additional neighbor_alltoall algorithms
The fix is for peioridic cartetian graph when dimension is 1 or 2. In
both case we have two edges points to the same process and vice versa.
If we send in the order of left and right, we need receive in the order
of right and left since left and right are reversed between in and out.

Previously we only fixed ineighbor_alltoall. This patch extends the same
fix to ineighbor_alltoallv and ineighbor_alltoallw.

I also updated the comment because the previous comments is still
confusing to me. In particular, it fails to high light the key points --
First it only matters for the periodic cartesian dimension of 1 or 2.
Second that results in multiple message with the same datatype and tag.
Lastly, the matching need reverse because left and right is reversed
between in and out.

Personally, I think allowing graphs to have duplicated edges pointing to
the same rank shouldn't be allowed. Because a general graph may not
guarantee ordering the edges and thus there is no way of ensuring the
correct matching order. In the general light, this fix is merely a hack.