posix: add release-gather based non-blocking collective framework
Implement non-blocking version of release_gather framework to implement
shm_ibcast and shm_ireduce using gentran scheduler. A schedule is created
in the form of a forest, where each tree represents a chunk of a large message.
Each chunk is given a unique sequence number so that all ranks know which
cell of shm buffer to use for copying the data in/out. Also, have a set
of release, gather flags per rank per cell of shm_buffer.