config: do not add flat_namespace option by default
We added -Wl,-flat_namespace when we used to depend on common symbols
to be shared across libmpifort.so and libmpi.so. This is no longer the
case. This flag is added to the mpicxx wrapper scripts and have side
effects on all dynamic libraries users may use. It is better we do not
add it by default.

I imagine if user use another Fortran library that built with MPI may
still need this option to make the constants, e.g. MPI_IN_PLACE, to
work.