debugger: Move queue pointers out of request union
The debug queue pointers are used for both send and receive request
types, and may be extended to others in the future. Place them outside
of the union in MPIR_Request to simplify access. Since these are
compiled out in a default build, the space needed is not a major
concern. Always initialize the queue pointers to NULL so that unexposed
requests are ignored at free time.