coll: Add multi-leader allreduce composition
Multi-leaders based composition: It has `num_leaders` per node,
which reduce the data within sub-node_comm. It is followed by
intra_node reduce and inter_node allreduce on the piece of data
the leader is responsible for. A shared memory buffer is
allocated per leader. If size of message exceeds this shm buffer,
the message is chunked.
Constraints: For a comm, all nodes should have same number
of ranks per node, op should be commutative.

Co-authored-by: Surabhi Jain <surabhi.jain@intel.com>