config: amend error messages for -fallow-argument-mismatch
While the main mpich configure only need FC and FCFLAGS, the configure
in test/mpi still need FFLAGS because there are tests in F77. Because
the test/mpi is a subdir within mpich configure, we need direct user to
supply both flags. This is not an issue in the main branch since we have
detached test/mpi from the mpich configure.