ch3: always pack low-density data
MPIR_Typerep_pack may leave off at non-iov boundary, which prevent
continuing loading iov. This is not an issue with dataloop but yaksa
does not allow it. Typically, if a datatype overall is high-density,
it won't benefit much from packing sporadic low-density area since there
won't be significant amount of low-density area. On the other hand, if
overall it is low-density, it is rare to have local high-density area.
Thus, simply use overall density should be sufficient for performance.
Doing so avoids mixing packing with using iovs.