fortran: add mismatch flags to mpif77 wrapper
Remove the AC_MSG_ERROR on needing the Fortran mismatch flag and add the
flag to the mpif77 wrapper.

This will not affect users of modern Fortran MPI applications who are
likely to use `mpifort` or `mpif90` and uses the mpi module interface.

Users who use `mpif77` is likely to use `mpif.h` interface and will
require the ignore-mismatch compiler flags to compile their MPI
applications.

By adding the extra flags to mpif77 wrapper only, we hope to strike a
balance of convenience and impact.