spawn: Avoid passing pointer to NULL info array
MPID_Comm_spawn_multiple takes an array of info ptrs or NULL as
argument. MPID_Comm_spawn calls MPID_Comm_spawn_multiple with count=1
and always passed an array of info pointers, even when the underlying
array did not exist. Check if there is an array, and pass NULL
otherwise.