configure: revamp options for pm and pmi
The current options are not ituitive and not easy for some options such
as with slurm either PMI1 or PMI2:
    * --with-pmi=simple --with-pm=hydra
    * --with-pmi=pmi2/simpl --with-pm=hydra
    * --with-pmi=cray --with-pm=no
    * --with-pmi=slurm --with-pm=no
    * --with-pmix=[path]

This commit makes 3 orthogonal options:
    * --with-pmi={pmi1, pmi2, pmix}
    * --with-pmilib={mpich, slurm, cray, pmix}
    * --with-pm={no,hydra,hydra2,gforker,remshell}

The --with-pm option is unchanged. The separation of --with-pmi and
--with-pmilib allows all possible combinations to be chosen.

We make it backward compatible with the old option such as --with-pmi=cray
by adding specific checks.

Extra complexities such as sourcing mpichprereq, sourcing pm_setup, are
removed, preferring explicit settings via case switches.