pmi2: remove reference of MPICH internals
Because the pmi implementation may be used without MPICH, we should not
refer to any MPICH internals. This commit removes references of -
  * mpiimpl.h
  * MPIR_Assert
  * MPIR_Memcpy
  * MPICH_IS_THREADED
  * MPIR_THREAD_CHECK_BEGIN

Because we need a way to tell the implementation whether thread
protection is required, this commit adds a necessary API -
  * PMI2_Set_threaded(int is_threaded)

Ideally, we should add parameters to PMI2_Init to pass the thread-level
requirement. Use a spearate API keeps backward compatibility.