pmi: add pmi_wire for handling PMI wire protocols
The wire protocols are othogonal to semantic protocols. The former
essentially is a serialization format. The latter denotes required or
expected key-values.

This commit add pmi_wire.h and pmi_wire.c to abstract the handling
of wire protocols. Currently PMI-1 and PMI-2 uses two different wire
protocols. But in principle, there is nothing prevents either wire
protocol to be used with either semantic protocols.

We define a PMIU_cmd object as an internal representation of a PMI
command. It can set/get/query its key-values and it can serialize to
different wire protocols.

Add PMIU_cmd_get_response to replace GetResponse in pmi_v1.c.