mpix: add MPIX_Stream
A new MPI handle object to represent a serial operation context. It
can be used to identify either a user thread, or a user GPU stream, or
any user-level serial context.

On creation, user set info hints to associate the newly created MPI
stream object to corresponding to their user-level objects, for example,
thread id or cudaStream_t. Accepted info keys are "type" and "id" (or
"value"). "type" can be "general" or "cudaStream_t" -- more to be added.

It is ok to create stream with MPI_INFO_NULL, in this general case
implementation understand the operations issued to this stream are
guaranteed to be serialized. Thus, using MPI Stream is more of a user
promise to allow potential implementation optimizations.

If a cuda stream is associated, any enqueue operation will be enqueued
to the cuda stream.