info: use dynamic array in MPIR_Info
The current code implements MPIR_Info as a linked list. It has a weird
property of keeping the first entry NULL that is not obvious and easy to
make mistake. Besides the code to iterate linked list is not as clean as
if we simply use a dynamic array.

The commit re-implements MPIR_Info using dynamic array.