hydra: remove forced_cleanup in proxy
forced_cleanup is used to reset the exit status of force killed
processes to 0. For example, when one process dies and auto_cleanup is
set, we kill the rest of the processes in the process group, and we want
to mark the exit status of those processes as 0. However, I don't think
that is necessary. Having the signal bit-or'ed in the final exit code
does no harm IMO. On the other hand, the extra logic is fragile. For
example, the presetting the exit code to 1 may mess up due to the order
of socket close event and waitpid event, and result in the temporary
exit code leaked to the final exit code and confuse users.