stream: use a single vci for gpu streams
It is often not critical for gpu streams to use separate vcis as long it
is separated from normal traffic. In addition, it can be common for an
application to use many GPU streams globally and we may not have enough
dedicated vcis to match them. This commit default to reuse a single
dedicated vci for gpu streams.