mpit: correctly initialize accum buffer for pvars
The PVAR SUM buffer was not initialized with zero after
allocation. As a result, the PVARs were getting unpredictable
initial values and that led to incorrect recording + reporting.

Apparently, in the past, we relied on calloc to initialize these
values. However, calloc has been replaced with malloc. So, that
expectation is not longer valid.