test/multinic: Fix pvar count assertion
The multinic counter pvars are arrays of MPI_LONG_LONG values. The count
value returned by MPI_T_pvar_handle_alloc is the size of the
array. ch4:ofi statically allocates an array of MPIDI_OFI_MAX_NICS,
which is currently defined to 8. In order for this test to accurately
check the expected values for each nic, the returned count should be
greater than or equal to the number of nics detected in the system.