errhan: split mpi-io related errhan routines
Split mpi-io related error handling routines to separate compile unit.
They reference routines defined in libromio.a but errhan_impl.c is
included in libpmpi.a, thus it is an issue in static linking. This patch
make sure the file errhan_file.c won't be referenced unless specific
MPI-IO error routines are used.

This patch won't address the issue of romio reference when these
routines are actually needed. That will be addressed separately.