ch4/ofi: Set MR coll_key_mask depending upon the max_mr_key_size
Previously, the requested key used in an fi_mr_reg call was using
an 8 byte value. This was causing failures for providers like cxi,
where the max allowed key value was 2 bytes. With this change,
the coll_key_mask is set to (max_mr_key_size in bits - 1)
instead of the default value of 63, allowing only valid values
as the requested key.