errhan: add comm_ptr to call_errhandler
Invoke MPIR_Handle_fatal_error on specific comm_ptr when it is
MPI_ERRORS_ABORT.

Note, this makes a slight difference between MPI_ERRORS_ARE_FATAL and
MPI_ERRORS_ABORT. The former in MPI_Comm_call_errhandler will abort the
whole world, which is the old behavior. However, the old behavior when
the error handler is triggered by actual errors is to abort on the
specific communicator, i.e. the same befavior per MPI_ERRORS_ABORT.

We are not changing the old behavior, thus the behavior of
MPI_ERRORS_ARE_FATAL is a bit inconsistent depend on whether it is
invoked explicitly or from error triggering.

We believe the intention of MPI_ERRORS_ABORT is to replace
MPI_ERRORS_ARE_FATAL due to its previous ambiguous semantics.