ch4: add MPIDI_REQ_TYPE
In MPID_Request_complete we need to do cleanups depend on the types of
request, in particular, whether it is a mpidig request. Currently we are
initializing and checking MPDIIG_REQUEST(req, req), a sub-field inside a
union. This usage is obscure and easy to miss. Instead, use a separate
type to tell what type of union the request is specialized into. We can
easily initialize the type to 0 and set accordingly when the request is
specialized.