hydra: fix hwloc paths and move them to modules dir
It is unnecessary to put embedded library to deep nested paths. Put them
all in modules dir, similar to how mpich does it.

The hwloc path was previously broken when we rearranged the source tree.
It didn't show in our testing because by default, we reuse the hwloc
compiled by main mpich.