ch4/gpu: insert ipc handles into the cache after creation
Insert into the cache immediately instead of waiting until fully
committing to the ipc path. Originally, if the base address met the
threshold but the user-defined length of the buffer did not, a new ipc
handle would be created. For the ze backend, this means a new opened fd.
By caching the handle immediately, we remove the risk of creating
spurious ipc handles.