coll: Add multileader alltoall composition
Author: Surabhi Jain <surabhi.jain@intel.com>

Leverage shared memory buffer to accumulate the data from ranks on a
same node. Create multi-leaders communicator among corresponding
ranks on a node. For example - a comm out of rank 0 of each node, a comm
out of rank 1 of each node and so on. Call inter-node Alltoall on these
multi-leader comms using the shm buffers of each node.