Docs: Python script to validate links for wiki
This program parses through all of the markdown files in the wiki/
directory and sub-directories. Web links are checked for a 200 HTTP
response and relative links are checked to see if the file exists.

Relative github links will not work (../this/file.c#L0-10).

Additionally the program will check for any files that are not linked to by any
other document, indicating they are not used anywhere.