ch4/ofi: Add cap set variable to determine support for null buffer registration
This commit adds MPIDI_OFI_ENABLE_MR_REGISTER_NULL, which indicates if the
provider supports registration of NULL buffers. This is useful for providers
like CXI which do not have FI_MR_ALLOCATED to be set, but still do not support
registration of null buffers. Currently, only the CXI provider has
MPIDI_OFI_ENABLE_MR_REGISTER_NULL set to OFF, with all other providers having
it as ON. If the CXI provider sets FI_MR_ALLOCATED, this
cap set variable can be removed in the future.