typerep: skip unsigned integer type for yaksa reduce
Yaksa currently treat unsigned integer type the same as its
corresponding signed type. This will not work in general. To fix, yaksa
need generate additional pup code for unsigned datatypes, which will
further bloating up the build time and resulting binary size.

Because yaksa reduce is an optimization path and unsigned integer
computation is not common in HPC applications, we opt for simply skip
yakda reduce in this case.