mpl: add interface for dev/subdev list query
One interface returns a list of dev_id strings or subdev_ids strings
for a given device. Another interface take a list of dev_id strings or
subdev_id strings and returns the environment variable value for
visibility control.

If any GPU visibility or affinity environment variable is define, the
return value of these functions must preserved the order defined in the
envs.