hydra: fix PMI2 kvs_fence
The fence_count need to be per pg or it will interfere between multiple
spawned groups.

Because the fence is non-blocking, we need trace epoch before we adding
up fence_count. When we reset fence_count, we need check and update the
count for all processes that are already in later epochs.