romio: actually set status in MPI_File_read_all
The old code blindly sets status of MPI_File_read_all with the input
data size. This ignore the cases when the actual file is too short or
have issues of actually getting the requested data. This patch updates
the status with actual number of bytes received.

We pass the status parameter to ADIOI_Read_and_exch and set the status
there. Add actual_recved_bytes parameter to ADIOI_R_Exchange_data so it
can return the actual bytes received in the buffer. Move ADIO_ReadContig
ahead before setting send_size so that we can set them reflecting the
actual amount of data we read.