mpl/gpu: add a function to check if a pointer is strictly a device allocation
For ZE, registered host and shared buffers are considered device
objects; passing one of those buffers to MPIR_GPU_query_pointer_is_dev
returns true. In some cases, we need to know if an allocation is
explicitly a device buffer. This is the purpose of
MPIR_GPU_query_pointer_is_strict_dev.