test/mpi: Remove runtime configure check for MPI_Publish_name
This check did not compile with a C99 compiler due to implicit
declaration of exit(), leading to name publishing tests being disabled
more often than they should. Additionally, a singleton test may not be
an accurate check, since publishing is often supported by the process
manager.

Remove this check and let the user enable/disable testing with the
provided switch or xfail infrastructure.