mpix: add MPIX_Type_iov and MPIX_Type_iov_len
MPIX_Type_iov_len returns the number of contiguous segments in a
datatype.

MPIX_Type_iov returns an array of MPIX_Iov at a given iov offset.

Each MPIX_Iov describes a contiguous segment. It is identical to struct
iovec. It is provided to avoid header dependency for mpi.h.

Internally this is equivalent to the corresponding typerep APIs with
buf address at 0 and count equal to 1.