typerep/dataloop: new algorithm for dataloop iov
Iterating segments to count number of contig segments is very slow.
Implement a recursive code similar to what we did in yaksa. The dataloop
are optimized, for example, reducing types into vector or contig when
segments are uniform or adjacent to each other. Thus, the dataloop iov
can be more effience than yaksa.