ch4/ucx: Only set window base pointer when allocating
We only need to update the window base pointer if we are using
ucp_mem_map to allocate the window memory. Otherwise we may overwrite
the actual base address with a different address, for instance if
ucp_mem_map chooses a page-aligned base pointer instead of the one
supplied as input. Fixes RMA test failures with UCX >= 1.13.0.