datatype: commit datatype in MPIR_Datatype_iscontig
The is_contig field is set after the type is committed. This allows the
typerep to optimize including whether the datatype is considered
contiguous, which needs to be consistent within typerep.

Thus we need make sure to commit the datatype before return the
is_contig.

MPIR_Datatype_iscontig is only used by ROMIO.