hydra/mpiexec: epoch_init after final process_count
In stead of calling HYD_pmiserv_epoch_init in
HYD_pmcd_pmi_alloc_pg_scratch, call it separately in
HYD_pmci_launch_procs and in spawn. This is because during spawn, the
process count is not known until we parsed all the exec list, but it's
convenient to have HYD_pmcd_pmi_alloc_pg_scratch called early so we can
add preput keys during parsing the execlist. This split allows for
better refactoring.