pmi: update pmi client using pmi_msg utilities
Factor and keep all the message protocol details, such as choice of key
names and required attributes, all in pmi_msg. The pmi_msg utilities
include handlers for both client and server, thus easier to keep the
protocols consistent.

Note that PMI2_Info_Get{Node,Job}Attr do not return error if the key is not
found. This is different behavior from PMI-1. We handle it in pmi_v2.c.