pmi: control the size of struct PMIU_cmd
Due to a few worst case such as spawn and mput, we use a large tokens
array in struct PMIU_cmd. Not only it wastes stack space, it may be
insufficient for extreme cases. Use a small static token buffer and
allow it fallback to dynamic allocation to address the concern. We'll
define a much larger MAX_PMI_ARGS once the tests clear.