hydra: fix singleton exit
The singleton init will have hydra_pmi_proxy hang in the loop of
HYDT_dmx_wait_for_event because the singleton's stdio are never
connected. Detect singleton case in finalize handler and set both the
out and err sockets to HYD_FD_CLOSED so that the main loop can exit.

We also need skip the collecting of exit_status because the singleton
process was never the child of the proxy process.