comm: Check for unmatched messages before destroying comm
If there are still messages using a communicator context_id in the
unexpected queue at destruction time, they could disturb message
matching if the context_id is reused. To prevent this, check for and
receive any unmatched messages during communicator destruction and
return an error to the user. These messages are likely the result of an
incorrect program, so notifying the user should help expose the issue.

Fixes pmodels/mpich#6184.