hydra: pass both id and pgid during initial handshake
During proxy initial handshake, it sends back pgid and proxy_id so that
the server can locate the proxy. This patch doesn't alter the
functionality but adds the pgid to init_hdr as a level of
redundancy. In later commits the pgid can be used for locate pg when we
use a single listen port for launching multiple proxies.