hydra/mpiexec: factor get_exit_status
We need collect the exit status at the end, but it is not obvious that
the pg pointer could be moved once we change the pg_list into a dynamic
list. Factor the get_exit_status routine out so we won't accidentally
use the same pg pointer after the event loop.

In addition, add notes as it is not obvious that we can't free the pg 0
even when it's processes all exited in the middle. As future TODO: we
should make the exit status a HYD_server_info global and set it when we
free the pg 0.