ch4: remove misleading comment
The comment doesn't make sense and probably should be deleted. The code
it commented on is just adding up local sizes. It is likely the
"allreduce" comment was meant on replacing the allgather of shared
table. But it appears the shared table need to be allgathered anyway.
The comment serves as a red herring.