hydra: ignore EINTR during write
Interrupt may happen during write to an fd, which is not a severe error
especially for stdout and stderr. This can easily happen a child process
exit and generate a SIGCHLD while some output message is generated. I am
not sure whether the message will get lost. We may have an issue when
interrupt happens during a PMI message and the write gets lost.
Hopefully this is not the case. But eitherway it won't be worse than
current code, which is to abort regardless.