hydra: check EAGAIN in read
When reading a hdr command with payload, it is possible to get EAGAIN between
reading the hdr and reading the payload. This happens easily when server
and proxy are on the same host. Check EAGAIN and retry when that occurs.

Remove the exception of STDIN_FILENO. A proxy may close its STDIN and
resulting one of its control fd coincide with STDIN_FILENO. Interrupt on
stdin to the server should be handled by signal handlers anyway.

Remove the comment since it is now redundant.