init: only do async progress under world init
With session init, the world part of initialization may not be ready and
async progress cannot be started. Previous commit (0913836) moved
MPII_init_async because we need initialize the global states so the
async progress can be launched on-demand by user. But we should check
the init state before we launch the ASYNC_PROGRESS.