gpu/ipc: add CVARs to select IPC protocol and engine type
MPIR_CVAR_CH4_IPC_GPU_READ_WRITE_PROCOL defines the protocol type for
IPC. The options are auto, read, or write. The read protocol opens the
IPC handle on the local device associated with the communication. The
write protocol opens the IPC handle on the remote device. Auto will
attempt to use the write protocol since it is typically higher
performance. If it cannot select the write protocol (i.e. the remote
device is not visible), then it will fallback to the read protocol.

MPIR_CVAR_CH4_IPC_GPU_ENGINE_TYPE defines the engine type used for all
IPC operations. The options are auto, compute, copy_high_bandwidth, and
copy_low_latency. Auto will attempt to use low-latency by default,
except when the remote device is not visible or the send and recv
devices are the same device. In that case, it will use the
copy_high_bandwidth engine.