coll: remove MPID_Send_coll etc and use attr instead
Remove the MPID_Send_coll and MPID_Isend_coll interface and use attr to
carry the errflag instead.

Also use attr to carry the errflag for MPIC_Ssend and MPIC_Issend.
Previously it sets the tag for error bits, which is not reliable. For
example, ch4:ofi does not use error bits in tag.