coll: fix ialltoallw inplace
In MPIR_Ialltoallw_intra_sched_inplace, we neglected to use the correct
index for getting recvtype_sz.

MPIR_TSP_Ialltoallw_sched_intra_inplace is not affected because it
send/recv the original datatypes. That is inefficient if the datatype is
large and non-contiguous. Wonder why it didn't do the same as
MPIR_Ialltoallw_intra_sched_inplace.