mpi.h: exclude portions of mpi.h when BUILD_MPI_ABI
When BUILD_MPI_ABI is defined, include mpi_abi_internal.h, and exclude
the portion that defines constants and typedefs that are already defined
in mpi_abi_internal.h. We essentially swapping the ABI for all internal
code.

Note that we need to always define the MPICH specific parts or some
internal code won't build.

MPICH_SUPPRESS_PROTOTYPES is no longer used, so we removed it.