ch4/shm: Add repeat address check for IPC path
If a send buffer under the IPC threshold is used more than once, then it
may be beneficial to map it anyway. Add in the capability to use IPC
even for small messages if the buffer is used more than once in the last
ten shm send operations. This feature can be disabled via CVAR.