mpio: use HAVE_ROMIO instead of MPI_MODE_RDONLY
It is more direct to check HAVE_ROMIO rather than check whether
MPI_MODE_RDONLY is defined as a side effect. In fact, the check is not
robust as MPI_MODE_XXX  constants can be defined as enums.

In src/binding/cxx/buildiface, we need include mpichconf.h first in
order to have the HAVE_ROMIO defined.