configure: deprecate the convenience options in --with-pmi
We allowed use of --with-pmi=[cray|slurm] as convenience options.
Deprecate these options to encourage downstream to use standardized PMI
interfaces.

Rename the cray option to oldcray since the newer Cray environment
should work with the standard `--with-pmi=path` or `--with-pmi2=path`
options directly.

The --with-pmilib option is used to configure special settings for slurm
and cray pmi. Deprecate this option as downstream, e.g. cray, is fixing
the non-conformant implementations.

User still can manually set --with-pmi in order to work with old Slurm
and Cray libraries until the newer implementation from Slurm and Cray
become ubiquitous.