ch4/ofi: remove eagain comm hints
Seems the usage of the "eagain" comm hint is for lightweight send to
fail after MPIR_CVAR_CH4_OFI_MAX_EAGAIN_RETRY. All the other paths are
hard-coded with unlimited retries. One reason for this design may be to
let user feed back on potential delay in lightweight send. But is very
difficult to utilize. Simplify the design by removing the "eagain" comm
hint and allow MPIR_CVAR_CH4_OFI_MAX_EAGAIN_RETRY to always limit the
number of retries in all paths.

The default is unlimited retries on -FI_EAGAIN and this commit does not
change that.