ch4/progress: yield on the VCI lock when doing progress
when doing progress on a VCI, we have to yield the lock.
Otherwise, it might causes a deadlock in case two threads progress on the same VCI.

This happened with two threads and two windows with PSCW.
The first thread on the target side was in MPI_Win_Wait hogging the lock of the VCI,
leaving the second thread then unable to send the AM message in MPI_Win_post.
On the origin side, one thread was stuck in MPI_Start, waiting for the
never coming Post-related message and the second thread stuck in the
MPI_Put, unable to acquire the lock.