coll: remove unnecessary setting tag in send request
The MPI_TAG field is cleared in send request because the collective path
may use tag to carry error bits. However, this is unnecessary since
we only call MPIR_Process_status for receive requests.

Writing to the status messes up the cacheline in the case of
pre-completed builtin send request.