threadcomm: order of modifying globals and thread_barrier
Need make sure we only modify global setting before the thread_barrier
in MPIX_Threadcomm_start and after thread_barrrier in
MPIX_Threadcomm_finish, and anything that depends on the global setting
on the other side of the thread_barrier.

NOTE: this includes malloc/free because when debug memtrace is on, it
will depend on MPIR_ThreadInfo.isThreaded.