threadcomm: modify isThreaded inside thread_barrier
When we repeated start and finish a threadcomm, there is a race
condition between earlier finish and the next start. Rather than
adding double thread_barrier, we can extend the thread_barrier to have
the critical code (that modifies global settings) inside the
thread_barrier between when every threads arrived and before the first
thread to leave.