ch4/coll: Use type extent in fallback check for reduce/allreduce
In release_gather implementations for reduce and allreduce, data is not
packed because we need to apply an operation to it. We need to fallback
when the extent is larger than the cell size. Fixes pmodels/mpich#6550.

TODO: pack reduction data in terms of basic type when possible. This
should make most noncontig reductions more efficient and utilize
release_gather collectives in more scenarios.